requirements.txt
keras==2.10.0
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==9.5.0
tensorflow==2.10.1
